+++ tsconfig.app.json
@@ -22,5 +22,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src", "Types", "Utils"]
+  "include": ["src", "Types", "Utils", "Layout"]
 }
